requirements.txt
fastapi
uvicorn
python-dotenv
requests
langchain-google-genai
